+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - fixed wood translation being missing
 - fixed sometimes removing stats from loot
 - fixed item rendering flickering in odd circumstances
+- fixed crash when smithing with an invalid material